Hotfix/amlogs credentials (#1)
* Create unique credential for each metric input

Create a credential based on the input name in order to support
multiple metric inputs.  Also, this hotfix will migrate the existing
credentials to new credentials to support upgrades.

* Create unique credential for each log input

Create a credential based on the input name in order to support
multiple activity and diagnostic log inputs. Also, this hotfix will
migrate the existing credentials to new credentials to support upgrades.

* Check for MSI use

App ID and key will be blank if using MSI creds, so only try to update
the storage passwords if not using MSI.